refactor: group and structure user and user-usage modules
