+++ src/libs/satellite/src/db/assert.rs
@@ -4,10 +4,12 @@ use crate::db::types::state::{DocAssertDelete, DocAssertSet, DocContext};
 use crate::errors::db::{JUNO_DATASTORE_ERROR_CANNOT_READ, JUNO_DATASTORE_ERROR_CANNOT_WRITE};
 use crate::hooks::{invoke_assert_delete_doc, invoke_assert_set_doc};
 use crate::types::store::StoreContext;
-use crate::usage::assert::{assert_user_usage_collection_data, increment_and_assert_db_usage};
-use crate::user::assert::{
+use crate::user::core::assert::{
     assert_user_collection_caller_key, assert_user_collection_data, assert_user_is_not_banned,
     assert_user_write_permission,
+};
+use crate::user::usage::assert::{
+    assert_user_usage_collection_data, increment_and_assert_db_usage,
 };
 use crate::{DelDoc, Doc, SetDoc};
 use candid::Principal;
