+++ src/libs/satellite/src/user/mod.rs
@@ -1,2 +1,2 @@
-pub mod assert;
-mod types;
+pub mod core;
+pub mod usage;
